event.js is getting a little hard to navigate.  split out the two implementations and move it to a subdir (event_api)
