+++ lib/event_api/index.test.js
@@ -1,15 +1,15 @@
 /* eslint-env node, jest */
 const path = require("path"),
-  event = require("./event"),
-  schema = require("./schema"),
-  tracker = require("./async_tracker"),
-  pkg = require(path.join(__dirname, "..", "package.json"));
+  event = require("."),
+  schema = require("../schema"),
+  tracker = require("../async_tracker"),
+  pkg = require(path.join(__dirname, "..", "..", "package.json"));
 
 beforeEach(() =>
   event.configure({ api: "libhoney-event", transmission: "mock", writeKey: "abc123" }));
-afterEach(() => event.resetForTesting());
+afterEach(() => event._resetForTesting());
 test("libhoney default config", () => {
-  const honey = event.apiForTesting().honey;
+  const honey = event._apiForTesting().honey;
   expect(honey.transmission.constructorArg.apiHost).toBe("https://api.honeycomb.io");
   expect(honey.transmission.constructorArg.dataset).toBe("nodejs");
   expect(honey.transmission.constructorArg.writeKey).toBe("abc123");
@@ -19,7 +19,7 @@ test("libhoney default config", () => {
 });
 
 test("startRequest starts tracking and creates an initial event, finishRequest sends it", () => {
-  const honey = event.apiForTesting().honey;
+  const honey = event._apiForTesting().honey;
   expect(tracker.getTracked()).toBeUndefined(); // context should be empty initially
 
   let requestPayload = event.startRequest({
@@ -56,7 +56,7 @@ test("startRequest starts tracking and creates an initial event, finishRequest s
   expect(postData[schema.TRACE_SPAN_NAME]).toBe("name");
   expect(postData[schema.TRACE_ID]).toBe(traceId);
   expect(postData[schema.TRACE_PARENT_ID]).toBeUndefined();
-  expect(postData[schema.TRACE_SPAN_ID]).toBe(traceId);
+  expect(postData[schema.TRACE_SPAN_ID]).not.toBeUndefined();
   expect(postData[schema.DURATION_MS]).not.toBeUndefined();
 });
 
@@ -139,7 +139,7 @@ test("sub-events can opt to rollup count/duration into request events", () => {
 });
 
 test("sub-events will get manual tracing fields", () => {
-  const honey = event.apiForTesting().honey;
+  const honey = event._apiForTesting().honey;
 
   let requestPayload = event.startRequest({
     [schema.EVENT_TYPE]: "source",
@@ -168,7 +168,7 @@ test("sub-events will get manual tracing fields", () => {
 
 describe("custom context", () => {
   test("it should be added to request events", () => {
-    const honey = event.apiForTesting().honey;
+    const honey = event._apiForTesting().honey;
 
     let requestPayload = event.startRequest({
       [schema.EVENT_TYPE]: "source",
@@ -185,7 +185,7 @@ describe("custom context", () => {
   });
 
   test("removing it works too", () => {
-    const honey = event.apiForTesting().honey;
+    const honey = event._apiForTesting().honey;
 
     let requestPayload = event.startRequest({
       [schema.EVENT_TYPE]: "source",
@@ -204,7 +204,7 @@ describe("custom context", () => {
   });
 
   test("it should be added to subevents sent after it was added", () => {
-    const honey = event.apiForTesting().honey;
+    const honey = event._apiForTesting().honey;
 
     let requestPayload = event.startRequest({
       [schema.EVENT_TYPE]: "source",
