+++ lib/event_api/libhoney.js
@@ -4,41 +4,12 @@ const libhoney = require("libhoney").default,
   process = require("process"),
   path = require("path"),
   uuidv4 = require("uuid/v4"),
-  tracker = require("./async_tracker"),
-  instrumentation = require("./instrumentation"),
-  DeterministicSampler = require("./deterministic_sampler"),
-  schema = require("./schema"),
-  pkg = require(path.join(__dirname, "..", "package.json")),
+  tracker = require("../async_tracker"),
+  instrumentation = require("../instrumentation"),
+  DeterministicSampler = require("../deterministic_sampler"),
+  schema = require("../schema"),
+  pkg = require(path.join(__dirname, "..", "..", "package.json")),
   debug = require("debug")(`${pkg.name}:event`);
-
-let eventAPI;
-
-function configure(opts = {}) {
-  if (eventAPI) {
-    return;
-  }
-
-  let api = LibhoneyEventAPI;
-  if (opts.api) {
-    if (opts.api === "libhoney-event") {
-      api = LibhoneyEventAPI;
-    } else if (opts.api === "mock") {
-      api = MockEventAPI;
-    } else {
-      throw new Error("Unrecognized .api.  valid options are 'libhoney-event' / 'mock'");
-    }
-  }
-  debug(`using api: ${api == LibhoneyEventAPI ? "libhoney-event" : "mock"}`);
-  eventAPI = new api(opts);
-}
-
-function apiForTesting() {
-  return eventAPI;
-}
-
-function resetForTesting() {
-  eventAPI = null;
-}
 
 const defaultName = "nodejs";
 
@@ -46,53 +17,7 @@ const incr = (payload, key, val = 1) => {
   payload[key] = (payload[key] || 0) + val;
 };
 
-class MockEventAPI {
-  constructor(opts) {
-    this.constructorArg = opts;
-    this.customContext = {};
-    this.eventStack = [];
-    this.sentEvents = [];
-    this.traceId = 0;
-  }
-
-  startRequest(metadataContext, traceId) {
-    let id = this.traceId++;
-    tracker.setTracked({ id: traceId || id });
-    return this.startEvent(metadataContext);
-  }
-  finishRequest(ev) {
-    this.finishEvent(ev);
-    tracker.deleteTracked();
-  }
-  startEvent(metadataContext) {
-    let context = tracker.getTracked();
-    const eventPayload = Object.assign({}, metadataContext, {
-      [schema.TRACE_ID]: context.id,
-      startTime: Date.now(),
-      startTimeHR: process.hrtime(),
-    });
-    this.eventStack.push(eventPayload);
-    return eventPayload;
-  }
-  finishEvent(ev, _rollup) {
-    Object.assign(ev, this.customContext, {
-      [schema.DURATION_MS]: 0,
-      [schema.BEELINE_VERSION]: pkg.version,
-    });
-    // pop it off the stack
-    const idx = this.eventStack.indexOf(ev);
-    this.eventStack = this.eventStack.slice(0, idx);
-    this.sentEvents.push(ev);
-  }
-  addContext(map) {
-    Object.assign(this.customContext, map);
-  }
-  removeContext(key) {
-    delete this.customContext[key];
-  }
-}
-
-class LibhoneyEventAPI {
+module.exports = class LibhoneyEventAPI {
   constructor(opts) {
     let sampleRate;
     let optsWithoutSampleRate = {};
@@ -168,23 +93,23 @@ class LibhoneyEventAPI {
     const context = tracker.getTracked();
     if (!context) {
       // valid, since we can end up in our instrumentation outside of requests we're tracking
-      askForIssue("no context in finishEvent.");
+      this.askForIssue("no context in finishEvent.");
       return;
     }
     if (context.stack.length === 0) {
       // this _really_ shouldn't happen.
-      askForIssue("no payload for event we're trying to finish (stack is empty).");
+      this.askForIssue("no payload for event we're trying to finish (stack is empty).");
       return;
     }
     const idx = context.stack.indexOf(ev);
     if (idx === -1) {
       // again, this _really_ shouldn't happen.
-      askForIssue("no payload for event we're trying to finish (event not found).");
+      this.askForIssue("no payload for event we're trying to finish (event not found).");
       return;
     }
     if (idx !== context.stack.length - 1) {
       // the event we're finishing isn't the most deeply nested one. warn the user.
-      askForIssue(
+      this.askForIssue(
         "finishing an event with unfinished nested events. almost certainly not what we want."
       );
     }
@@ -263,50 +188,15 @@ class LibhoneyEventAPI {
       .concat(context.stack.map((payload, idx) => `${idx}: ${JSON.stringify(payload)}`))
       .join("\n");
   }
-}
 
-const customContext = {
-  add(key, val) {
-    eventAPI.addContext({ [schema.customContext(key)]: val });
-  },
-  remove(key) {
-    eventAPI.removeContext(schema.customContext(key));
-  },
+  askForIssue(msg, logger = debug) {
+    logger(`-------------------
+    ${pkg.name} error: ${msg}
+    please paste this message (everything between the "----" lines) into an issue
+    at ${pkg.bugs}.  feel free to edit
+    out any application stack frames if you'd rather not share those
+    ${new Error().stack}
+    ${this.dumpRequestContext()}
+    -------------------`);
+  }
 };
-
-exports.configure = configure;
-exports.apiForTesting = apiForTesting;
-exports.resetForTesting = resetForTesting;
-exports.traceActive = function traceActive() {
-  return !!tracker.getTracked();
-};
-exports.startRequest = function startRequest(metadataContext, traceId) {
-  return eventAPI.startRequest(metadataContext, traceId);
-};
-exports.finishRequest = function finishRequest(ev) {
-  return eventAPI.finishRequest(ev);
-};
-exports.addContext = function addContext(map) {
-  return eventAPI.addContext(map);
-};
-exports.startEvent = function startEvent(metadataContext, spanId) {
-  return eventAPI.startEvent(metadataContext, spanId);
-};
-exports.finishEvent = function finishEvent(ev, rollup) {
-  return eventAPI.finishEvent(ev, rollup);
-};
-const dumpRequestContext = (exports.dumpRequestContext = function dumpRequestContext() {
-  return eventAPI.dumpRequestContext();
-});
-exports.customContext = customContext;
-
-const askForIssue = (exports.askForIssue = (msg, logger = debug) => {
-  logger(`-------------------
-${pkg.name} error: ${msg}
-please paste this message (everything between the "----" lines) into an issue
-at ${pkg.bugs}.  feel free to edit
-out any application stack frames if you'd rather not share those
-${new Error().stack}
-${dumpRequestContext()}
--------------------`);
-});
