+++ lib/index.js
@@ -1,5 +1,5 @@
 /* eslint-env node */
-const event = require("./event"),
+const event = require("./event_api"),
   instrumentation = require("./instrumentation");
 
 function configure(opts = {}) {
