+++ lib/instrumentation/child_process.js
@@ -1,6 +1,6 @@
 /* eslint-env node */
 const shimmer = require("shimmer"),
-  event = require("../event"),
+  event = require("../event_api"),
   schema = require("../schema");
 
 function wrapExecLike(name, packageVersion) {
