+++ lib/instrumentation/express.js
@@ -2,7 +2,7 @@
 const onHeaders = require("on-headers"),
   tracker = require("../async_tracker"),
   schema = require("../schema"),
-  event = require("../event"),
+  event = require("../event_api"),
   path = require("path"),
   pkg = require(path.join(__dirname, "..", "..", "package.json")),
   debug = require("debug")(`${pkg.name}:express`);
