+++ lib/instrumentation/express.test.js
@@ -3,7 +3,7 @@ const request = require("supertest"),
   path = require("path"),
   instrumentExpress = require("./express"),
   schema = require("../schema"),
-  event = require("../event"),
+  event = require("../event_api"),
   pkg = require(path.join(__dirname, "..", "..", "package.json"));
 
 const tests = [
@@ -44,7 +44,7 @@ for (let t of tests) {
       initializeTestServer();
     });
     afterEach(() => {
-      event.resetForTesting();
+      event._resetForTesting();
       server.close();
     });
 
@@ -53,8 +53,8 @@ for (let t of tests) {
       request(server)
         .get("/")
         .expect(200, () => {
-          expect(event.apiForTesting().sentEvents.length).toBe(1);
-          let ev = event.apiForTesting().sentEvents[0];
+          expect(event._apiForTesting().sentEvents.length).toBe(1);
+          let ev = event._apiForTesting().sentEvents[0];
           expect(ev.startTime).not.toBeUndefined();
           expect(ev.startTimeHR).not.toBeUndefined();
           delete ev.startTime;
@@ -83,9 +83,8 @@ for (let t of tests) {
             "request.user.id": 42,
             "request.user.username": "toshok",
             "response.status_code": "200",
-            appEventPrefix: undefined,
           });
-          event.apiForTesting().sentEvents.splice(0, 1);
+          event._apiForTesting().sentEvents.splice(0, 1);
           done();
         });
     });
@@ -118,7 +117,7 @@ describe("userContext as function", () => {
     initializeTestServer();
   });
   afterEach(() => {
-    event.resetForTesting();
+    event._resetForTesting();
     server.close();
   });
 
@@ -127,8 +126,8 @@ describe("userContext as function", () => {
       .get("/")
       .expect(200, () => {
         expect(cb_called).toBe(true);
-        expect(event.apiForTesting().sentEvents.length).toBe(1);
-        let ev = event.apiForTesting().sentEvents[0];
+        expect(event._apiForTesting().sentEvents.length).toBe(1);
+        let ev = event._apiForTesting().sentEvents[0];
         expect(ev["request.user.random"]).toBe("stuff");
         done();
       });
@@ -154,7 +153,7 @@ describe("request id from http headers", () => {
     initializeTestServer();
   });
   afterEach(() => {
-    event.resetForTesting();
+    event._resetForTesting();
     server.close();
   });
 
@@ -163,8 +162,8 @@ describe("request id from http headers", () => {
       .get("/")
       .set("X-Request-ID", "abc123")
       .expect(200, () => {
-        expect(event.apiForTesting().sentEvents.length).toBe(1);
-        let ev = event.apiForTesting().sentEvents[0];
+        expect(event._apiForTesting().sentEvents.length).toBe(1);
+        let ev = event._apiForTesting().sentEvents[0];
         expect(ev[schema.TRACE_ID]).toBe("abc123");
         expect(ev[schema.TRACE_ID_SOURCE]).toBe("X-Request-ID http header");
         done();
@@ -176,8 +175,8 @@ describe("request id from http headers", () => {
       .get("/")
       .set("X-Amzn-Trace-Id", "Root=1-67891233-abcdef012345678912345678")
       .expect(200, () => {
-        expect(event.apiForTesting().sentEvents.length).toBe(1);
-        let ev = event.apiForTesting().sentEvents[0];
+        expect(event._apiForTesting().sentEvents.length).toBe(1);
+        let ev = event._apiForTesting().sentEvents[0];
         expect(ev[schema.TRACE_ID]).toBe("Root=1-67891233-abcdef012345678912345678");
         expect(ev[schema.TRACE_ID_SOURCE]).toBe("X-Amzn-Trace-Id http header");
         done();
@@ -190,8 +189,8 @@ describe("request id from http headers", () => {
       .set("X-Request-ID", "abc123")
       .set("X-Amzn-Trace-Id", "Root=1-67891233-abcdef012345678912345678")
       .expect(200, () => {
-        expect(event.apiForTesting().sentEvents.length).toBe(1);
-        let ev = event.apiForTesting().sentEvents[0];
+        expect(event._apiForTesting().sentEvents.length).toBe(1);
+        let ev = event._apiForTesting().sentEvents[0];
         expect(ev[schema.TRACE_ID]).toBe("abc123");
         expect(ev[schema.TRACE_ID_SOURCE]).toBe("X-Request-ID http header");
         done();
@@ -220,7 +219,7 @@ describe("trace id callback", () => {
     initializeTestServer();
   });
   afterEach(() => {
-    event.resetForTesting();
+    event._resetForTesting();
     server.close();
   });
 
@@ -229,8 +228,8 @@ describe("trace id callback", () => {
       .get("/")
       .set("X-Request-ID", "abc123")
       .expect(200, () => {
-        expect(event.apiForTesting().sentEvents.length).toBe(1);
-        let ev = event.apiForTesting().sentEvents[0];
+        expect(event._apiForTesting().sentEvents.length).toBe(1);
+        let ev = event._apiForTesting().sentEvents[0];
         expect(ev[schema.TRACE_ID]).toBe("abc123");
         expect(ev[schema.TRACE_ID_SOURCE]).toBe("traceIdSource function");
         done();
@@ -241,8 +240,8 @@ describe("trace id callback", () => {
     request(server)
       .get("/")
       .expect(200, () => {
-        expect(event.apiForTesting().sentEvents.length).toBe(1);
-        let ev = event.apiForTesting().sentEvents[0];
+        expect(event._apiForTesting().sentEvents.length).toBe(1);
+        let ev = event._apiForTesting().sentEvents[0];
         expect(ev[schema.TRACE_ID]).toBe("efgh456");
         expect(ev[schema.TRACE_ID_SOURCE]).toBe("traceIdSource function");
         done();
