+++ lib/instrumentation/http.js
@@ -1,7 +1,7 @@
 /* eslint-env node */
 const url = require("url"),
   shimmer = require("shimmer"),
-  event = require("../event"),
+  event = require("../event_api"),
   schema = require("../schema");
 
 let instrumentHTTP = (http, opts = {}) => {
@@ -21,11 +21,18 @@ let instrumentHTTP = (http, opts = {}) => {
         return original.apply(this, [options, cb]);
       }
 
+      // let's make sure we have an object (or a url) for options.
+      if (typeof options === "string") {
+        options = url.parse(options);
+      } else {
+        options = Object.assign({}, options);
+      }
+
       let ev = event.startEvent({
         [schema.EVENT_TYPE]: "http",
         [schema.PACKAGE_VERSION]: opts.packageVersion,
-        [schema.TRACE_SPAN_NAME]: "request",
-        url: typeof options === "string" ? options : url.format(options),
+        [schema.TRACE_SPAN_NAME]: options.method || "GET",
+        url: url.format(options),
       });
 
       let wrapped_cb = function(res) {
