+++ lib/instrumentation/http.test.js
@@ -3,7 +3,7 @@ const http = require("http"),
   instrumentHttp = require("./http"),
   schema = require("../schema"),
   tracker = require("../async_tracker"),
-  event = require("../event");
+  event = require("../event_api");
 
 instrumentHttp(http);
 
@@ -23,16 +23,15 @@ afterAll(() => {
 });
 
 test("GET http://localhost:9009/ works", done => {
-  let context = {};
+  let context = { stack: [] };
   tracker.setTracked(context);
 
   http.get("http://localhost:9009", _res => {
-    expect(tracker.getTracked()).toBe(context);
-    expect(event.apiForTesting().sentEvents).toMatchObject([
+    expect(event._apiForTesting().sentEvents).toMatchObject([
       {
         [schema.EVENT_TYPE]: "http",
-        [schema.TRACE_SPAN_NAME]: "request",
-        url: "http://localhost:9009",
+        [schema.TRACE_SPAN_NAME]: "GET",
+        url: "http://localhost:9009/",
       },
     ]);
     done();
