+++ lib/instrumentation/https.js
@@ -1,7 +1,7 @@
 /* eslint-env node */
 const url = require("url"),
   shimmer = require("shimmer"),
-  event = require("../event"),
+  event = require("../event_api"),
   schema = require("../schema");
 
 let instrumentHTTPS = (https, opts = {}) => {
@@ -21,11 +21,18 @@ let instrumentHTTPS = (https, opts = {}) => {
         return original.apply(this, [options, cb]);
       }
 
+      // let's make sure we have an object (or a url) for options.
+      if (typeof options === "string") {
+        options = url.parse(options);
+      } else {
+        options = Object.assign({}, options);
+      }
+
       let ev = event.startEvent({
         [schema.EVENT_TYPE]: "https",
         [schema.PACKAGE_VERSION]: opts.packageVersion,
-        [schema.TRACE_SPAN_NAME]: "request",
-        url: typeof options === "string" ? options : url.format(options),
+        [schema.TRACE_SPAN_NAME]: options.method || "GET",
+        url: url.format(options),
       });
 
       let wrapped_cb = function(res) {
