+++ lib/instrumentation/https.test.js
@@ -3,7 +3,7 @@ const https = require("https"),
   instrumentHttps = require("./https"),
   schema = require("../schema"),
   tracker = require("../async_tracker"),
-  event = require("../event");
+  event = require("../event_api");
 
 instrumentHttps(https);
 
@@ -14,16 +14,16 @@ beforeAll(() => {
 });
 
 test("GET https://google.com works", done => {
-  let context = {};
+  let context = { stack: [] };
   tracker.setTracked(context);
 
   https.get("https://google.com", _res => {
     expect(tracker.getTracked()).toBe(context);
-    expect(event.apiForTesting().sentEvents).toMatchObject([
+    expect(event._apiForTesting().sentEvents).toMatchObject([
       {
         [schema.EVENT_TYPE]: "https",
-        [schema.TRACE_SPAN_NAME]: "request",
-        url: "https://google.com",
+        [schema.TRACE_SPAN_NAME]: "GET",
+        url: "https://google.com/",
       },
     ]);
     done();
