+++ lib/instrumentation/mongodb.js
@@ -1,7 +1,7 @@
 /* eslint-env node */
 const shimmer = require("shimmer"),
   tracker = require("../async_tracker"),
-  event = require("../event"),
+  event = require("../event_api"),
   schema = require("../schema");
 
 function prefixKeys(prefix, map = {}) {
