+++ lib/instrumentation/mysql2.js
@@ -1,7 +1,7 @@
 /* eslint-env node */
 const shimmer = require("shimmer"),
   tracker = require("../async_tracker"),
-  event = require("../event"),
+  event = require("../event_api"),
   schema = require("../schema");
 
 let instrumentConnection = function(conn, packageVersion) {
