+++ lib/instrumentation/react-dom-server.test.js
@@ -4,7 +4,7 @@ const ReactDOMServer = require("react-dom/server"),
   instrumentReactDOMServer = require("./react-dom-server"),
   schema = require("../schema"),
   tracker = require("../async_tracker"),
-  event = require("../event");
+  event = require("../event_api");
 
 instrumentReactDOMServer(ReactDOMServer);
 
@@ -14,56 +14,56 @@ beforeAll(() => {
 
 describe("renderToString", () => {
   test("works with context (and sends event)", () => {
-    let context = {};
+    let context = { stack: [] };
     tracker.setTracked(context);
 
     ReactDOMServer.renderToString(React.createElement("div"));
 
-    expect(event.apiForTesting().sentEvents).toEqual([
+    expect(event._apiForTesting().sentEvents).toEqual([
       expect.objectContaining({
         [schema.EVENT_TYPE]: "react",
         [schema.TRACE_SPAN_NAME]: "renderToString",
         [schema.DURATION_MS]: expect.any(Number),
       }),
     ]);
-    expect(event.apiForTesting().sentEvents.length).toBe(1);
+    expect(event._apiForTesting().sentEvents.length).toBe(1);
     // XXX more here
-    event.apiForTesting().sentEvents.splice(0, 1);
+    event._apiForTesting().sentEvents.splice(0, 1);
   });
 
   test("works without context (sends no event)", () => {
     tracker.runWithoutTracking(() => {
       ReactDOMServer.renderToString(React.createElement("div"));
 
-      expect(event.apiForTesting().sentEvents.length).toBe(0);
+      expect(event._apiForTesting().sentEvents.length).toBe(0);
     });
   });
 });
 
 describe("renderToStaticMarkup", () => {
   test("works with context (and sends event)", () => {
-    let context = {};
+    let context = { stack: [] };
     tracker.setTracked(context);
 
     ReactDOMServer.renderToStaticMarkup(React.createElement("div"));
 
-    expect(event.apiForTesting().sentEvents).toEqual([
+    expect(event._apiForTesting().sentEvents).toEqual([
       expect.objectContaining({
         [schema.EVENT_TYPE]: "react",
         [schema.TRACE_SPAN_NAME]: "renderToStaticMarkup",
         [schema.DURATION_MS]: expect.any(Number),
       }),
     ]);
-    expect(event.apiForTesting().sentEvents.length).toBe(1);
+    expect(event._apiForTesting().sentEvents.length).toBe(1);
     // XXX more here
-    event.apiForTesting().sentEvents.splice(0, 1);
+    event._apiForTesting().sentEvents.splice(0, 1);
   });
 
   test("works without context (sends no event)", () => {
     tracker.runWithoutTracking(() => {
       ReactDOMServer.renderToStaticMarkup(React.createElement("div"));
 
-      expect(event.apiForTesting().sentEvents.length).toBe(0);
+      expect(event._apiForTesting().sentEvents.length).toBe(0);
     });
   });
 });
